requirements.txt
# Core dependencies for diffusion models
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.6.0
pillow>=9.2.0
numpy>=1.22.0
tqdm>=4.64.0

# Diffusers and transformers
diffusers>=0.18.0,<0.20.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.1
datasets>=2.14.0  # For HuggingFace datasets (QuickDraw)

# Jupyter notebook support
ipython>=8.10.0
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: for high quality image processing
opencv-python>=4.6.0